Moved transformers below models
* Moved the transformers to below models, they are transformers for model data.
