+++ app/Http/Controllers/ItemSubCategoryController.php
@@ -6,7 +6,7 @@ use App\Http\Parameters\Route\Validate;
 use App\Models\ItemCategory;
 use App\Models\ItemSubCategory;
 use App\Models\SubCategory;
-use App\Transformers\ItemSubCategory as ItemSubCategoryTransformer;
+use App\Models\Transformers\ItemSubCategory as ItemSubCategoryTransformer;
 use App\Utilities\Request as UtilityRequest;
 use App\Http\Parameters\Request\Validators\ItemSubCategory as ItemSubCategoryValidator;
 use Exception;
