+++ app/Models/Transformers/Category.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Transformers;
+namespace App\Models\Transformers;
 
 use App\Models\Category as CategoryModel;
 use App\Models\SubCategory as SubCategoryModel;
