+++ app/Models/Transformers/ItemMonthSummary.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Transformers;
+namespace App\Models\Transformers;
 
 use App\Models\Item as ItemModel;
 
@@ -12,28 +12,28 @@ use App\Models\Item as ItemModel;
  * @copyright Dean Blackborough 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ItemYearSummary extends Transformer
+class ItemMonthSummary extends Transformer
 {
-    private $year_summary;
+    private $month_summary;
 
     /**
      * ResourceType constructor.
      *
-     * @param ItemModel $year_summary
+     * @param ItemModel $month_summary
      */
-    public function __construct(ItemModel $year_summary)
+    public function __construct(ItemModel $month_summary)
     {
         parent::__construct();
 
-        $this->year_summary = $year_summary;
+        $this->month_summary = $month_summary;
     }
 
     public function toArray(): array
     {
         return [
-            'id' => $this->year_summary->year,
-            'year' => $this->year_summary->year,
-            'total' => number_format((float) $this->year_summary->total, 2, '.', '')
+            'id' => $this->month_summary->month,
+            'month' => date("F", mktime(0, 0, 0, $this->month_summary->month, 1)),
+            'total' => number_format((float) $this->month_summary->total, 2, '.', '')
         ];
     }
 }
